+++ libnopegl/src/hmap.h
@@ -22,6 +22,8 @@
 #ifndef HMAP_H
 #define HMAP_H
 
+#include <stdlib.h>
+
 #ifndef HMAP_SIZE_NBIT
 #define HMAP_SIZE_NBIT 3
 #endif
@@ -45,7 +47,7 @@ typedef void (*user_free_func_type)(void *user_arg, void *data);
 
 struct hmap *ngli_hmap_create(void);
 void ngli_hmap_set_free(struct hmap *hm, user_free_func_type user_free_func, void *user_arg);
-int ngli_hmap_count(const struct hmap *hm);
+size_t ngli_hmap_count(const struct hmap *hm);
 int ngli_hmap_set(struct hmap *hm, const char *key, void *data);
 void *ngli_hmap_get(const struct hmap *hm, const char *key);
 struct hmap_entry *ngli_hmap_next(const struct hmap *hm, const struct hmap_entry *prev);
